requirements/development.txt
# python -m venv venv
# python -m pip install --upgrade pip setuptools
# pip install -r requirements\development.txt --upgrade

-r base.txt

Werkzeug
django-debug-toolbar
pylint
pylint-django
# django-autotranslate     # Language translation